DNA-278 Create new utility class for determining validity of XML NCNames
Created the implementation class by using a single structure to hold for each character a bitmask that describes the set of classifications that character belongs to.  For example, 'a' belongs to multiple character classes, so the bitmask ORs together the bitmask of each of those character classes.  A simple array is used to store all the bitmasks, indexed by the integer value of the character.  Then, each method (e.g., to check whether a character is a valid NCName starting character) simply has to look up the mask and perform a single bit operation.  (Some methods also have some shortcuts, as several of the classes have a very large range of characters (at the upper end) that are or are not in the character class, and these can be checked by simply comparing the integer value of the character with the lower end of the range.)

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@719 76366958-4244-0410-ad5e-bbfabb93f86b