DNA-281 Connector support for workspaces
Implemented the changes to support workspaces in connectors.  This involved a lot of changes, including all connectors, the Graph API (and implementation), and all uses of the graph API (including a lot of test cases).

The SimpleRepositorySource that was available in dna-graph/src/java/test was removed because it started to mirror the code that is in the InMemoryRepositorySource (which is now in dna-graph).  This affected quite a few tests cases.

This change also made apparent a couple of issues with the current federation code.  These problems require more in-depth changes and will be done next, but shouldn't be mixed with this set of changes.  This also has implications for several higher-level projects (dna-jcr, getting started examples) that use the connector.  Therefore, these projects were temporarily removed and will be added back in shortly.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@723 76366958-4244-0410-ad5e-bbfabb93f86b