DNA-283 fileSystemPaths property value is ignored for FileSystemSources configured through XML
Changed RepositoryService to do a better job at figuring out how to convert the property values in the configuration repository into value(s) that are acceptable to the setter method. This uses the value factories to do the conversion, so RepositoryService really only works with RepositorySource implementations that have property setters that accept any of our known/supported property types or primitives (or arrays of those property types or primitives).

In order to do this, I added a new method added to PropertyType to discover the property type given a Class parameter.


git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@724 76366958-4244-0410-ad5e-bbfabb93f86b