DNA-280 Create optimized versions of Location
Applied the patch to change from a constructor-based Location to a static factory method style.  This will allow us to create specialized subclasses, optimized for the different combinations of attributes.  Using a static factory method isn't much different than using a constructor - the caller still refers to the class at compile-time.  The factory methods are not there to allow someone else to replace or specialize the Location; it is there so that we can hide which of our implementations we want to use.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@731 76366958-4244-0410-ad5e-bbfabb93f86b