DNA-286 Implement namespace management
This change provides a new implementation of the NamespaceRegistry interface of the JCR API, along with refactoring the implementations of DNA's NamespaceRegistry.  In particular, a new GraphNamespaceRegistry class was added that stores the namespace mappings in the graph content (under a location provided into the constructor).  This allowed the new JcrNamespaceRegistry to simply delegate to that and do the simple mapping from DNA's interface to JCR's.  (The refactoring also changed a number of test cases that were directly instantiating a namespace registry.)

The other major change to the JCR implementation is the proper implementation of a session-specific (non-shared and non-persistent) namespace registry, on top of the shared (in a content-sense, not shared objects) and persistent registry that is in the JcrWorkspace.  This required changing how the JcrSession and JcrWorkspace objects are created.  The result is that now namespace mappings can be set on the session, and these override those in the workspace (that would be used for all sessions that use that repository's workspace).

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@738 76366958-4244-0410-ad5e-bbfabb93f86b