DNA-286 Implement namespace management
This change corrects the behavior of the Session's namespace-related methods, which are not the same as those in the NamespaceRegistry implementation associated with the workspace.  This also corrected several of the TCK tests associated with namespaces.

Note that the StringValueFactory in 'dna-graph' was changed to use a DNA namespace registry to convert a Name and Path to a string using the namespace prefixes.  (This change affected a number of unit tests.)  The result is that any Name and Path property values are treated correctly and properly converted to a string using the current namespace prefixes in the namespace registry.  However, the BinaryValueFactory implementations should not use this StringValueFactory, since they need to convert values using the non-prefixed form (that is, the namespace URI).

In the case of our JCR implementation, the above change to the Graph factories means that property values are properly mapped to strings using the Session's local mappings (which fall back on the Workspace's mappings).  This corrected at least one TCK test.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@743 76366958-4244-0410-ad5e-bbfabb93f86b