DNA-290 Getting child nodes or properties by pattern is not implemented
Implemented the AbstractJcrNode.getNodes(String pattern) and AbstractJcrNode.getProperties(String pattern), which makes a few more of the JCR TCK tests pass.  In the process, the way AbstractJcrNode stores children was (dramatically) simplified, and a few different kinds of NodeIterator and PropertyIterator implementations were added (including empty ones).

Also changed how DNA Path.Segment objects are created when no SNS index is specified, so the objects are now created with an index of 1 rather than -1.  This was done because it was leaking -1 in some cases, and it also more closely aligns with JCR.  Found and corrected minor related bugs in a few test cases.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@750 76366958-4244-0410-ad5e-bbfabb93f86b