DNA-293 JcrSession creates Property instances that use the PropertyType.UNDEFINED, which is not allowed
Changed JcrSession.populateNode(...) to never use PropertyType.UNDEFINED for the actual property type on a property.  This means that even though UNDEFINED is allows on PropertyDefinition, it is not allowed on an actual Property instance.  Therefore, I added the actual property type as an int field to JcrAbstractProperty, and modified the subclasses and the code where the constructors are called (mostly JcrSession and unit tests).

The current choice is to use PropertyType.STRING, since all property values can be converted to a String.  See the issue for more details and a discussion of this logic.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@755 76366958-4244-0410-ad5e-bbfabb93f86b