DNA-194 Implement update JCR capability
None of the nodes had a Node.getDefinition() implementation.  To properly implement this, I refactored how JcrSession was creating and populating nodes, and I added some utility methods to the JcrNodeTypeManager (which is now a separate package-protected class rather than an inner class).

This allows all of the Level 2 TCK unit tests to get further, but none of the other update functionality is implemented, so none of the Level 2 TCK tests run successfully.

I also made a number of utility methods final, in the hopes that they can be inlined since they are called very frequently.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@761 76366958-4244-0410-ad5e-bbfabb93f86b