DNA-194 Implement update JCR capability
Applied the "...move_value_cast" patch, which moves the JcrNodeType logic determining whether a value can be cast to another type into the JcrValue class where it can be used to actually convert the value to a value of another type.  Also made the "createValueFormatException(...)" methods private to increase likelihood they'll be inlined. All tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@770 76366958-4244-0410-ad5e-bbfabb93f86b