DNA-194 Implement update JCR capability
Added most of the infrastructure to allow updates to the SessionCache.  The cached representations of the nodes are changed and corresponding requests are accumulated in a Graph.Batch that are sent to the repository store when SessionCache.save() is called.  This implementation hooks some of the infrastructure up to the JcrAbstractNode implementation, although more tests need to be written to verify the behavior.  

Currently, only deleting nodes has been tested and is thought to be complete.  Changing properties, creating children, and moving nodes are partially implemented.  (Some changes to the node type manager are required, but there is currently an ongoing refactoring of JcrNodeTypeManager.  This is the primary reason for committing this partially-complete work.)


git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@784 76366958-4244-0410-ad5e-bbfabb93f86b