DNA-302 Make JcrNodeTypeManager not dependent on JcrSession
Applied the patch, which splits JcrNodeTypeManager into repository-scoped functionality (RepositoryNodeTypeManager) and session-scoped functionality (JcrNodeTypeManager) with the key difference being that the session-scoped node type manager is able to translate names based on transient namespace remappings. 

The RepositoryNodeTypeManager is the maintainer of node type information for the entire repository at run-time. The 
repository manager maintains a list of all node types and the ability to retrieve node types by Name. 

The JCR 1.0 and 2.0 specifications both require that node type information be shared across all sessions within a 
repository and that the javax.jcr.NodeTypeManager perform operations based on the string versions of Names based 
on the permanent (workspace-scoped) and transient (session-scoped) namespace mappings. DNA achieves this by 
maintaining a single master repository of all node type information (the RepositoryNodeTypeManager) and per-session 
wrappers (JcrNodeTypeManager) for this master repository that perform String to Name translation based on the 
javax.jcr.Session's transient mappings and then delegating node type lookups to the repository manager. 

Also removed JcrSession references from all the type-related objects and replaced them with ExecutionContexts.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@785 76366958-4244-0410-ad5e-bbfabb93f86b