DNA-333 BasicName.getString Doesn't Use Provided TextEncoder When the Name Has No Namespace
Applied the patch that corrects BasicName to use the encoder if suppliedto getString(), specifically when the name contains no namespace (e.g., "*").  The patch also changes the behavior of BasicName.getString(TextEncoder) to return the encoded local name ONLY when called. It used to return "{}" + localName.

Verified that all tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@787 76366958-4244-0410-ad5e-bbfabb93f86b