DNA-305 JcrNodeType.canSetProperty Does Not Enforce Constraints
Applied the "DNA-305_path_namespace" patch that makes path constraints support transient namespace remapping, modifies the test cases to verify this, and moves the test node type into class to clean up the package and make JcrPropertyDefinitionTest more consistent with the approach in SessionCacheTest. This fix should allow the defect to be resolved.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@790 76366958-4244-0410-ad5e-bbfabb93f86b