DNA-335 DNA's NodeTypeManager implementation does not allow multiple definitions with the same name
Applied the patch ("DNA-335_clean_type_source") that simplifies RepositoryTypeManager's internal type handling by combining the primaryNodeTypes and mixinNodeTypes into a single map under the assumption that lookups of types by name (which must check primary and mixin types) will be the most common option and retrieval of all primary or mixin types will be relatively rare. 

This also allows for a simplification of the JcrNodeTypeSource hierarchy.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@791 76366958-4244-0410-ad5e-bbfabb93f86b