DNA-306 Add ability to run all TCK unit tests within Hudson, showing success and failure rate
Applied the patch, which adds a 'dna-jcr-tck' project (which is no longer under 'extensions' but is at the top level) that shows how the project is doing relative to the JCR TCK unit tests.  These unit tests are NOT the official TCK, but are reportedly some of the tests that comprise the TCK and are available via the Jackrabbit project.  Also changed the parent POM to include this module only when using the 'jcr-tck' profile (e.g., 'mvn install -P jcr-tck').

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@792 76366958-4244-0410-ad5e-bbfabb93f86b