DNA-335 DNA's NodeTypeManager implementation does not allow multiple definitions with the same name
Applied the "DNA-335_overriding_definitions" patch that changes the behavior to be more consistent with the implications of section 4.7.7 in the JCR 2.0 spec. Subtypes can override definitions of their supertypes (or their super-supertypes, etc.). Included some very rudimentary test cases as a sanity check. A more complete test suite is forthcoming. 

This patch retains the support for multiple inheritance but adds the assumption that any future type registration support will limit overriding to cases where the overriding property or child node definition is "more narrow" than the definition being overridden. In the case of properties, this implies that the required type of the property in the subtype's property definition can ALWAYS be cast to the required type of the property in the supertype's definition.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@793 76366958-4244-0410-ad5e-bbfabb93f86b