DNA-336 JcrValue.asType(int desiredPropertyType) skips a number of valid conversions and has duplicate logic
Changed the JcrValue.asType(int) logic to first ensure the value matches the current required type (in most cases this should just return the actual value if the type is correct), and to then convert to the desired property type.  Corrected a number of conversions to include several combinations that was excluded.

Also added a check to the NameValueFactory to ensure that the relative path (of size 1) has no SNS index.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@794 76366958-4244-0410-ad5e-bbfabb93f86b