DNA-340 Projection of Node Types onto System View
Applied the patch, which stores the node type information in the repository graph, with one exception.  For various reasons (including those related to performance of running so many TCK unit tests), we want to be able to have a JcrRepository configuration property that says whether the node types should be stored in the graph.  So, the "batch.execute()" line was commented out at the end of the RepositoryNodeTypeManager.projectOnto(Graph,Path) method.  This essentially prevents the node types from being added, but still ensures that the "/jcr:system/dna:nodeTypes" node exists.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@804 76366958-4244-0410-ad5e-bbfabb93f86b