DNA-340 Projection of Node Types onto System View
Applied the 2nd patch, which adds support to JcrRepository for configurable options, with a single option as to whether or not the node types should be projected under "/jcr:system/dna:nodeTypes".  Minor change from the patch was to reorder the constructor parameters of JcrRepository to match the other existing constructor.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@805 76366958-4244-0410-ad5e-bbfabb93f86b