DNA-341 Graph API behavior when removing all property values doesn't reflect that of JCR
Applied patch that changes the connector API.  Specifically the semantics and a few method signatures of UpdatePropertiesRequest class were changed to no longer assume that an empty property should be removed; rather, the names of properties that are to be removed are tracked explicitly, and the number of values in a property no longer has any relationship to whether the property is to be removed.  All connectors were changed to apply these new semantics.

A number of other changes to the connector API were made.  New SetPropertyRequest and RemovePropertyRequest requests types were added as being more efficient when a single property is being set or removed.  These are optional requests, meaning that the RequestProcessor has default implementations for the corresponding "process" methods that convert the reqeusts to UpdatedPropertiesRequest.  Thus, a connector doesn't have to implement these, but it may be more efficient for it to do so.  The RemovedPropertiesRequest was also removed, since it didn't offer any benefit over UpdatePropertiesRequest (and since it is simpler to only have one).  All connectors were updated to reflect these semantics.

Also, the Graph batch implementation was improved to optimize adjacent calls to set/remove properties on the same node.  In other words, if a batch is used to set a property on node A and then is immediately used to set another property on the same node, the batch will combine the two requests.  However, this optimization does not apply if any other operations are performed between the set/remove property operations.

Finally, a minor error was found and corrected in how the connector unit tests reported the (trivial) metrics for creating nodes. The total times (line 2) were correct, but the average and the times for additional nodes (line 3) were all incorrect.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@809 76366958-4244-0410-ad5e-bbfabb93f86b