DNA-339 JCR Type Registration and Unregistration
Applying the patch, with some very minor modifications (particularly, some additional JavaDoc in NodeDefinitionId and PropertyDefinitionId).  The Key classes in JcrPropertyDefinition and JcrNodeDefinition should be replaced with PropertyDefinitionId and NodeDefinitionId, respectively, but this can be done in another change set.  The current approach works, but may actually not match the specification.  (The JavaDoc added to the *DefinitionId classes helps to explain what information exactly is used to distinguish the definitions, and the Key classes use a different set of information.)

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@820 76366958-4244-0410-ad5e-bbfabb93f86b