DNA-354 JcrSession.hasPendingChanges Is Not Implemented
Applied the patch (with changes) that implements this method by checking the SessionCache for unsaved (pending) changes.  The patch had checked the 'changes' map to determine whether there were changes, but this misses deletes.  Added a method to SessionCache that determines whether there are any pending changes, and implemented this method by checking whether there are any pending operations.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@826 76366958-4244-0410-ad5e-bbfabb93f86b