DNA-354 JcrSession.hasPendingChanges Is Not Implemented
Previous changes that were made in addition to the patch did not properly implement the hasPendingChanges method.  When doing this, it was discovered that SessionCache was incorrectly creating it's Graph.Batch instance to not use the SessionCache's request builder (which used the cache's list of requests, allowing SessionCache.save(UUID) to play with the pending requests to save only those that affect the branch below the supplied UUID).  

This commit corrects the implementation, and all tests now pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@827 76366958-4244-0410-ad5e-bbfabb93f86b