DNA-356 Need to Throw ItemExistsException if addNode Call Fails Due to SNS
Added a patch that changes SessionCache.NodeEditor.addNode(...) to throw an ItemExistsException if the node cannot be added because an item at the specified path already exists and same-name siblings are not allowed, per the specification (see the JIRA issue for the details).  An additional lookup of the child node definition is performed, but this isn't a problem since it is already an error case and the extra work is justified to provide more useful information in the exception.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@830 76366958-4244-0410-ad5e-bbfabb93f86b