DNA-357 Hide dna:nodeDefinition Property from Users
Applied both patches.  The first hides the "dna:nodeDefinition" property from the JCR layer (meaning it might exist in the graph layer, but won't appear as properties of JCR nodes).  The second patch adds to "nt:base" a property definition for a "dna:nodeDefinition" property that is protected and not mandatory (nor autocreated).  The spec is not clear as to whether a JCR implementation can add properties to "nt:base" (the required supertype of all node types), the spec is clear that an implementation can add properties to any of the built-in node types through mixins.  Therefore, the second patch actually adds this new definition to "nt:base" by having "nt:base" extend the new "dna:defined" mixin, which has the new "dna:nodeDefinition" property definition.  

Because the "dna:nodeDefinition" is protected, JCR clients will not be allowed to define another property with this name, nor will they be able to set the property (using a residual definition) to any value.  All this validation is done through the normal process, requiring no code specific to "dna:nodeDefinition".  Nicely done!

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@831 76366958-4244-0410-ad5e-bbfabb93f86b