DNA-359 Session.refresh and Node.refresh Methods Not Supported
Applied the patch that implements the Session.refresh and Node.refresh by adding two SessionCache.refresh methods (one to refresh all contents, the other to refresh only those items below a branch).  

There is one case that the patch doesn't handle (well), and that is when a node is removed from a branch, refreshing that branch will not refresh the removed node.  This needs to be addressed, and will be described in the issue.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@832 76366958-4244-0410-ad5e-bbfabb93f86b