DNA-363 Define node types via CND files
The first step of several.  This change moves some of the names in the JCR-related lexicons from the "dna-jcr" project's lexicons into the "dna-graph" project's corresponding lexicons.  

Also added to the "nt:nodeType" and "nt:propertyDefinition" node types in JcrBuiltinNodeTypeSource were several additional property definitions new in JSR-283 .  However, because the JSR-170 TCK unit tests are checking that only the JSR-170 property definitions are there, these new definitions have been commented out for the time being.  (JSR-283 also makes all of the property definitions on "nt:nodeType", "nt:propertyDefinition", and "nt:childNodeDefinition" to be protected, though JSR-170 TCK unit tests are checking that they are not.  These also were left as not being protected so that we are JSR-170 compliant.)

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@836 76366958-4244-0410-ad5e-bbfabb93f86b