DNA-58 Create repository configuration and component
Refactored the DnaConfiguration to make use of the graph repository for storing the configuration.  When used, the DnaConfiguration instance persists all the information to this graph, where it can be read and used.  A "save()" method was added to allow the code using the DnaConfiguration instance to control when the content is updated.  (This also simplified the unit testing, since it's very easy to check the graph content after the DnaConfiguration methods are used.)  The DnaEngine was then updated to read the configuration repository content and to instantiate and set up the services.

This "configuration repository" defaults to an in-memory repository, but can be set up to use a particular RepositorySource instance (that has already been set up) or to reflectively create an instance given the class name, optional class loader, and various (implementation-specific) Java bean properties.  

This means that in the simplest case, one can simply create a DnaConfiguration and configure the sequencers, sources, detectors, etc.  It also makes it very simple to "read" an XML file with the configuration, since that just uses the XML graph importer.  But using a configuration repository shines in the more advanced cases, such as when a process participates in a cluster.  Here, the process can create a DnaConfiguration, very easily set it up to use a shared (and remote) configuration repository (that already has the content), and without any other work immediately build the DnaEngine.

Using a graph as the configuration repository also means that any changes to the configuration could be monitored via graph events, allowing the DnaEngine to monitor any changes and dynamically alter its services.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@849 76366958-4244-0410-ad5e-bbfabb93f86b