DNA-58 Create repository configuration and component
Changed DnaConfiguration to make it more tolernant of changes to the source pointing to the configuration repository.  In particular, this allows the DnaEngine to obtain an immutable definition of the configuration source, which allows the engine to properly access the source even if the DnaConfiguration instance is changed.

The DnaEngine was also refactored to pull into a nested scanner class the code that reads the configuration content.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@850 76366958-4244-0410-ad5e-bbfabb93f86b