DNA-374 Multiplicity of some properties may change when saved and reloaded from the graph
Added a patch that fixes this issue by maintaining a hidden property to store the names of the properties that use multi-valued definitions but that have only one value.  This property is persisted as required, but is also cleaned up if it is empty or no longer needed.

The patch also moves the "dna:nodeDefinition" and this new property into a new "dnaint" namespace.  This allows some of the code to just check the namespace to know whether properties should be hidden from the JCR view.

Committed the patch, which also added in a number of commented-out TCK unit tests.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@857 76366958-4244-0410-ad5e-bbfabb93f86b