DNA-387 Federation connector does not support updates, which are required when using with the JCR implementation
Added support for writes in limited scenarios (when the nodes being changed are within a single projection), and addressed a couple of issues related to the handling of UUIDs in the federation connector.  Added more unit tests, verifying the behavior is correct.  Also corrected a few errors in CopyBranchRequest, MoveBranchRequest, and CreateNodeRequest.  Finally, the repository examples (from the Getting Started document) were corrected to use our JCR implementation.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@866 76366958-4244-0410-ad5e-bbfabb93f86b