DNA-58 Create repository configuration and component
Refactored the DnaConfiguration class to use more interfaces and to move those interfaces (and reusable implementations) into a Configurator class that uses generics.  DnaConfiguration instantiates a configurator, but also implements some of the high-level interfaces by delegating them to the wrapped configurator instance.  This allows DnaConfiguration to appear to be doing the configurating, but in reality it is the wrapped builder that is doing it.  Also, DnaConfiguration (or other components) can choose which "aspects" of configuration it is to expose by inheriting controlling which higher-level interfaces are implemented.  Since delegation is used, this works very well.

This is a more complicated design, but it is far more extensible and reusable.  A JcrConfiguration class was also created that works in the same way as DnaConfiguration, except that it uses a different type parameter (so all the interfaces return JcrConfiguration).  It thus gets to reuse all of the Configurator logic.  However, JcrConfiguration doesn't implement all of the "aspects", essentially hiding certain services (e.g., sequencers, since they don't yet work with our implementation).

This change also made a few minor corrections to the JavaDoc and visibility modifiers on the classes in the 'org.jboss.dna.jcr' package.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@867 76366958-4244-0410-ad5e-bbfabb93f86b