DNA-277 Fix all text files in source code to have consistent line feeds
Used Eclipse to convert line endings to Unix, which removes the CR and forces all files to have a LF line terminator.  Also added a README_COMMITTERS.txt file that specifies what should be added to a developer's subversion configuration to ensure files are always uploaded with the correct terminators.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@869 76366958-4244-0410-ad5e-bbfabb93f86b