DNA-385 Documentation Updates for 0.4
Reviewed and edited both documents to bring everything up to the current state of the code.  Most parts were still accurate, but there were a number of places there were minor changes and quite a few sections that needed more work.

Pending errors or other changes/suggestions, the docs are complete and accurate for this release.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@870 76366958-4244-0410-ad5e-bbfabb93f86b