DNA-394 Maven reports Java sequencer project is missing indirectly referenced Eclipse OSGI JAR
Changed the referenced version of the org.eclipse:osgi artifact from a range to the specific version that is in the JBoss repository, and moved the dependency fragment ahead of all other Eclipse dependencies.  This appears to work, though I'm still not sure which Eclipse artifact was depending upon this version such that when the range was used, the versions clashed or differed.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@885 76366958-4244-0410-ad5e-bbfabb93f86b