DNA-398 RepositoryNodeTypeManager is Not Thread-Safe
Corrected the public getAllNodeTypes() method, which was returning the actually values collection associated with an internal map that is modified.  This means that although the method atomically returns the collection in a thread-safe manner, the resulting collection is live and may be modified via other calls.  The change is to create a defensive and immutable copy of the node types within the lock scope.  Currently, this method is only called from the JcrNodeTypeManager implementation (the Session-specific wrapper around the shared RepositoryNodeTypeManager), which is not currently called from within any DNA JCR implementation code.  Therefore, the cost of making a copy has little impact.

Also added were JavaDoc for several public methods.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@897 76366958-4244-0410-ad5e-bbfabb93f86b