DNA-402 Update MIME type based on file extensions
The previous mechanism for specifying the extensions for MIME types was a property file. This format is not the conventional way systems typically have this information formatted; usually it is a *nix-style tab-delimited file with the mime type followed by whitespace (usually tab) followed by any applicable extensions.  Therefore, the property-file approach was changed to support these *nix-style files directly, though property file formats are still allowed (delimited by '=' rather than whitespace).  A regular expression is used to process each line, so the code works with both file formats.  Changed the name of the resource/classpath file used for the defaults from MimeTypes.properties to mime.types, and changed this file accordingly.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@902 76366958-4244-0410-ad5e-bbfabb93f86b