DNA-403 Federation connector does not behave properly in many situations
Implemented a new federation connector (in 'dna-graph') that addresses most if not all of the concerns listed in the issue.  Some changes were made to the requests (e.g., addition of DeleteChildrenRequest) and ExecutionContext and Location.  However, most of the changes were in the addition of the 'org.jboss.dna.graph.connector.federation' package and its mostly package-level classes.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@911 76366958-4244-0410-ad5e-bbfabb93f86b